[core] -Fix: hid_attrib: copy HATT_ENUM as (long *) because the target field is really a long
git-svn-id: svn://repo.hu/librnd/trunk@32562 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/librnd/core/hid_attrib.c
@@ -134,7 +134,7 @@ int rnd_hid_parse_command_line(int *argc, char ***argv)
 				if ((backup != NULL) && (a->value == NULL))
 					a->value = &backup[i].lng;
 				if (a->value)
-					*(int *) a->value = a->default_val.lng;
+					*(long *) a->value = a->default_val.lng;
 				break;
 			case RND_HATT_UNIT:
 				if ((backup != NULL) && (a->value == NULL))
